using multiprocessing methods better
